requirements.txt
annotated-types==0.7.0
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1733250440834/work
axial-attention==0.6.1
Brotli @ file:///C:/b/abs_c415aux9ra/croot/brotli-split_1736182803933/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1746569525376/work/certifi
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.8
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
colorlog==6.9.0
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1733502965406/work
contourpy==1.3.1
cramjam==2.9.1
cycler==0.12.1
debugpy @ file:///D:/bld/debugpy_1744321248910/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1740384970518/work
docker-pycreds==0.4.0
einops==0.8.1
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1746947292760/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1745502089858/work
fastparquet==2024.11.0
filelock @ file:///C:/b/abs_f2gie28u58/croot/filelock_1700591233643/work
fonttools==4.57.0
fsspec==2025.3.2
fvcore==0.1.5.post20221221
gitdb==4.0.12
GitPython==3.1.44
gmpy2 @ file:///C:/b/abs_d8ki0o0h97/croot/gmpy2_1738085498525/work
GPUtil==1.4.0
grad-cam==1.5.5
greenlet==3.2.2
huggingface-hub==0.30.2
idna @ file:///C:/b/abs_aad84bnnw5/croot/idna_1714398896795/work
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1737420181517/work
importlib_resources==6.5.2
intel-openmp==2021.4.0
iopath==0.1.10
ipykernel @ file:///D:/bld/ipykernel_1719845595208/work
ipython @ file:///D:/bld/bld/rattler-build_ipython_1745673220/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1733300866624/work
Jinja2 @ file:///C:/b/abs_920kup4e6u/croot/jinja2_1741711580669/work
joblib==1.4.2
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1733440914442/work
jupyter_core @ file:///D:/bld/jupyter_core_1727163532151/work
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe @ file:///C:/b/abs_a0ma7ge0jc/croot/markupsafe_1738584052792/work
matplotlib==3.10.1
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1733416936468/work
mdurl==0.1.2
MedPy==0.5.2
mkl==2021.4.0
mkl-fft==1.3.1
mkl-random @ file:///C:/ci_310/mkl_random_1643050563308/work
mkl-service==2.4.0
mpmath @ file:///C:/b/abs_7833jrbiox/croot/mpmath_1690848321154/work
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1733325553580/work
networkx @ file:///C:/b/abs_b054htfn9t/croot/networkx_1737043671910/work
nibabel==5.3.2
numpy @ file:///C:/b/abs_54abayvc9j/croot/numpy_and_numpy_base_1682520598361/work
opencv-python==4.11.0.86
packaging==25.0
pandas==2.2.3
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1733271261340/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1733327343728/work
pillow==11.1.0
platformdirs==4.3.7
portalocker==3.1.1
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1744724089886/work
protobuf==5.29.4
psutil @ file:///D:/bld/psutil_1740663160591/work
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1733569405015/work
pyarrow==19.0.1
pydantic==2.11.2
pydantic_core==2.33.1
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1736243443484/work
pyparsing==3.2.3
PySocks @ file:///C:/ci_310/pysocks_1642089375450/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1733215673016/work
pytz==2025.2
pywin32==307
PyYAML @ file:///C:/b/abs_14xkfs39bx/croot/pyyaml_1728657968772/work
pyzmq @ file:///D:/bld/pyzmq_1743831231478/work
regex==2024.11.6
requests @ file:///C:/b/abs_c3508vg8ez/croot/requests_1731000584867/work
rich==14.0.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
segment-anything==1.0
sentry-sdk==2.25.1
setproctitle==1.3.5
simpleitk==2.5.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.2
SQLAlchemy==2.0.40
sqlmodel==0.0.24
stack_data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1733569443808/work
sympy @ file:///C:/b/abs_b4u17p23yg/croot/sympy_1738108511395/work
tabulate==0.9.0
tbb==2021.13.1
termcolor==3.0.1
threadpoolctl==3.6.0
timm==1.0.15
tokenizers==0.21.1
toml==0.10.2
torch==2.3.0
torchaudio==2.3.0
torchsummary==1.5.1
torchvision==0.18.0
tornado @ file:///D:/bld/tornado_1732615925919/work
tqdm==4.67.1
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1733367359838/work
transformers==4.51.3
ttach==0.0.3
typing-inspection==0.4.0
typing_extensions @ file:///C:/b/abs_0ffjxtihug/croot/typing_extensions_1734714875646/work
tzdata==2025.2
urllib3 @ file:///C:/b/abs_7bst06lizn/croot/urllib3_1737133657081/work
wandb==0.19.9
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1733231326287/work
win-inet-pton @ file:///C:/ci_310/win_inet_pton_1642658466512/work
yacs==0.1.8
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1732827521216/work
